Tests: Move composite fixtures to conftest
- move composite fixtures to conftest such that they can be used by
  other tests, and to conform to the 'fixtures in conftest.py' structure
- rename fixture 'composite_model' to 'model_vision'
- add alexnet to the model_vision fixture
