+++ tests/test_composites.py
@@ -2,18 +2,7 @@
 from types import MethodType
 from itertools import product
 
-import pytest
-from torchvision.models import vgg11, resnet18
-
 from zennit.core import BasicHook, collect_leaves
-from zennit.composites import COMPOSITES, NameMapComposite, LayerMapComposite, SpecialFirstLayerMapComposite
-from zennit.composites import EpsilonGammaBox
-
-
-@pytest.fixture(scope='session', params=[vgg11, resnet18])
-def composite_model(request):
-    '''Models to test composites on.'''
-    return request.param()
 
 
 def ishookcopy(hook, hook_template):
@@ -51,82 +40,45 @@ def verify_no_hooks(model):
     )
 
 
-SPECIAL_FIRST_LAYER_MAP_COMPOSITES = [
-    elem for elem in COMPOSITES.values() if issubclass(elem, SpecialFirstLayerMapComposite)
-]
-LAYER_MAP_COMPOSITES = [
-    elem for elem in COMPOSITES.values()
-    if issubclass(elem, LayerMapComposite) and not issubclass(elem, SpecialFirstLayerMapComposite)
-]
-COMPOSITE_KWARGS = {
-    EpsilonGammaBox: {'low': -3., 'high': 3.},
-}
-
-
-@pytest.fixture(scope='session', params=LAYER_MAP_COMPOSITES)
-def layer_map_composite(request):
-    '''Fixture for explicit LayerMapComposites.'''
-    return request.param(**COMPOSITE_KWARGS.get(request.param, {}))
-
-
-@pytest.fixture(scope='session', params=SPECIAL_FIRST_LAYER_MAP_COMPOSITES)
-def special_first_layer_map_composite(request):
-    '''Fixturer for explicit SpecialFirstLayerMapComposites.'''
-    return request.param(**COMPOSITE_KWARGS.get(request.param, {}))
-
-
-@pytest.fixture(scope='session')
-def name_map_composite(request, composite_model, layer_map_composite):
-    '''Fixture to create NameMapComposites based on explicit LayerMapComposites.'''
-    rule_map = {}
-    for name, child in composite_model.named_modules():
-        for dtype, hook_template in layer_map_composite.layer_map:
-            if isinstance(child, dtype):
-                rule_map.setdefault(hook_template, []).append(name)
-                break
-    name_map = [(tuple(value), key) for key, value in rule_map.items()]
-    return NameMapComposite(name_map=name_map)
-
-
-def test_composite_layer_map_registered(layer_map_composite, composite_model):
+def test_composite_layer_map_registered(layer_map_composite, model_vision):
     '''Tests whether the explicit LayerMapComposites register and unregister their rules correctly.'''
     errors = []
-    with layer_map_composite.context(composite_model):
-        for child, (dtype, hook_template) in product(composite_model.modules(), layer_map_composite.layer_map):
+    with layer_map_composite.context(model_vision):
+        for child, (dtype, hook_template) in product(model_vision.modules(), layer_map_composite.layer_map):
             if isinstance(child, dtype) and not check_hook_registered(child, hook_template):
                 errors.append((
                     '{} is of {} but {} is not registered!',
                     (child, dtype, hook_template),
                 ))
 
-    if not verify_no_hooks(composite_model):
+    if not verify_no_hooks(model_vision):
         errors.append(('Model has hooks registered after composite was removed!', ()))
 
     assert not errors, 'Errors:\n  ' + '\n  '.join(f'[{n}] ' + msg.format(*arg) for n, (msg, arg) in enumerate(errors))
 
 
-def test_composite_special_first_layer_map_registered(special_first_layer_map_composite, composite_model):
+def test_composite_special_first_layer_map_registered(special_first_layer_map_composite, model_vision):
     '''Tests whether the explicit LayerMapComposites register and unregister their rules correctly.'''
     errors = []
     try:
         special_first_layer, special_first_template, special_first_dtype = next(
             (child, hook_template, dtype)
             for child, (dtype, hook_template) in product(
-                collect_leaves(composite_model), special_first_layer_map_composite.first_map
+                collect_leaves(model_vision), special_first_layer_map_composite.first_map
             ) if isinstance(child, dtype)
         )
     except StopIteration:
         special_first_layer = None
         special_first_template = None
 
-    with special_first_layer_map_composite.context(composite_model):
+    with special_first_layer_map_composite.context(model_vision):
         if special_first_layer is not None and not check_hook_registered(special_first_layer, special_first_template):
             errors.append((
                 'Special first layer {} is of {} but {} is not registered!',
                 (special_first_layer, special_first_dtype, special_first_template)
             ))
 
-        children = (child for child in composite_model.modules() if child is not special_first_layer)
+        children = (child for child in model_vision.modules() if child is not special_first_layer)
         for child, (dtype, hook_template) in product(children, special_first_layer_map_composite.layer_map):
             if isinstance(child, dtype) and not check_hook_registered(child, hook_template):
                 errors.append((
@@ -134,17 +86,17 @@ def test_composite_special_first_layer_map_registered(special_first_layer_map_co
                     (child, dtype, hook_template),
                 ))
 
-    if not verify_no_hooks(composite_model):
+    if not verify_no_hooks(model_vision):
         errors.append(('Model has hooks registered after composite was removed!', ()))
 
     assert not errors, 'Errors:\n  ' + '\n  '.join(f'[{n}] ' + msg.format(*arg) for n, (msg, arg) in enumerate(errors))
 
 
-def test_composite_name_map_registered(name_map_composite, composite_model):
+def test_composite_name_map_registered(name_map_composite, model_vision):
     '''Tests whether the constructed NameMapComposites register and unregister their rules correctly.'''
     errors = []
-    setups = product(composite_model.named_modules(), name_map_composite.name_map)
-    with name_map_composite.context(composite_model):
+    setups = product(model_vision.named_modules(), name_map_composite.name_map)
+    with name_map_composite.context(model_vision):
         for (name, child), (names, hook_template) in setups:
             if name in names and not check_hook_registered(child, hook_template):
                 errors.append(
@@ -152,7 +104,7 @@ def test_composite_name_map_registered(name_map_composite, composite_model):
                     (name, hook_template),
                 )
 
-    if not verify_no_hooks(composite_model):
+    if not verify_no_hooks(model_vision):
         errors.append(('Model has hooks registered after composite was removed!', ()))
 
     assert not errors, 'Errors:\n  ' + '\n  '.join(f'[{n}] ' + msg.format(*arg) for n, (msg, arg) in enumerate(errors))
